+++ lib/ui/widgets/dropdown.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:zachranobed/common/constants.dart';
+import 'package:zachranobed/common/utils/field_validation_utils.dart';
 
 class ZODropdown extends StatefulWidget {
   final String hintText;
@@ -24,26 +25,46 @@ class ZODropdown extends StatefulWidget {
 }
 
 class _ZODropdownState extends State<ZODropdown> {
+  bool _isValid = true;
+
   @override
   Widget build(BuildContext context) {
-    return DropdownButtonFormField(
-      isExpanded: true,
-      hint: Text(widget.hintText),
-      validator: widget.onValidation,
-      items: widget.items.map<DropdownMenuItem<String>>((String value) {
-        return DropdownMenuItem<String>(value: value, child: Text(value));
-      }).toList(),
-      onChanged: (String? value) {
-        setState(() {
-          widget.onChanged(value!);
-        });
+    return TapRegion(
+      onTapOutside: (tap) {
+        // Force remove focus when dropdown menu is collapsed
+        FocusScope.of(context).requestFocus(FocusNode());
       },
-      value: widget.initialValue,
-      decoration: const InputDecoration(
-        enabledBorder: WidgetStyle.inputBorder,
-        focusedBorder: WidgetStyle.inputBorder,
+      child: DropdownButtonFormField(
+        isExpanded: true,
+        validator: FieldValidationUtils.wrapValidator(
+          widget.onValidation,
+          (isValid) {
+            setState(() {
+              _isValid = isValid;
+            });
+          },
+        ),
+        items: widget.items.map<DropdownMenuItem<String>>((String value) {
+          return DropdownMenuItem<String>(
+            value: value,
+            child: Text(
+              value,
+              style: Theme.of(context).textTheme.bodyLarge,
+            ),
+          );
+        }).toList(),
+        onChanged: (String? value) {
+          setState(() {
+            widget.onChanged(value!);
+          });
+        },
+        value: widget.initialValue,
+        decoration: WidgetStyle.createInputDecoration(
+          label: widget.hintText,
+          isValid: _isValid,
+        ),
+        focusNode: widget.focusNode,
       ),
-      focusNode: widget.focusNode,
     );
   }
 }